+++ kyuubi-ha/src/main/scala/org/apache/kyuubi/ha/HighAvailabilityConf.scala
@@ -149,8 +149,8 @@ object HighAvailabilityConf {
       .booleanConf
       .createWithDefault(false)
 
-  val HA_ZK_ENGINE_SECRET_NODE: OptionalConfigEntry[String] =
-    buildConf("ha.zookeeper.engine.secret.node")
+  val HA_ZK_ENGINE_SECURE_SECRET_NODE: OptionalConfigEntry[String] =
+    buildConf("ha.zookeeper.engine.secure.secret.node")
       .doc("The zk node contains the secret that used for internal secure between Kyuubi server " +
         "and Kyuubi engine, please make sure that it is only visible for Kyuubi.")
       .version("1.5.0")
