+++ kyuubi-ha/src/main/scala/org/apache/kyuubi/service/authentication/ZooKeeperEngineSecureSecretProviderImpl.scala
@@ -20,7 +20,7 @@ package org.apache.kyuubi.service.authentication
 import java.nio.charset.StandardCharsets
 
 import org.apache.kyuubi.config.KyuubiConf
-import org.apache.kyuubi.ha.HighAvailabilityConf.HA_ZK_ENGINE_SECRET_NODE
+import org.apache.kyuubi.ha.HighAvailabilityConf.HA_ZK_ENGINE_SECURE_SECRET_NODE
 import org.apache.kyuubi.ha.client.ZooKeeperClientProvider
 
 class ZooKeeperEngineSecureSecretProviderImpl extends EngineSecureSecretProvider {
@@ -34,7 +34,7 @@ class ZooKeeperEngineSecureSecretProviderImpl extends EngineSecureSecretProvider
 
   override def getSecret(): String = {
     var secret: String = ""
-    conf.get(HA_ZK_ENGINE_SECRET_NODE).map { zkNode =>
+    conf.get(HA_ZK_ENGINE_SECURE_SECRET_NODE).map { zkNode =>
       withZkClient(conf) { zkClient =>
         secret = new String(zkClient.getData.forPath(zkNode), StandardCharsets.UTF_8)
       }
